# Introduce multi-layered canvas rendering for optimization