--- a/index.html
+++ b/index.html
@@ -34,7 +34,10 @@
       <h5 id="game-over">GAME OVER</h5>
       <div class="screen">
         <img id="youngImg" src="res/young.png"/>
-        <canvas id="canvas" width="500" height="680"></canvas>
+        <div id="stage">
+            <canvas id="canvas" width="500" height="680"></canvas>
+            <canvas id="background" width="500" height="680"></canvas>
+        </div>
         <img id="chanImg" src="res/chan.png"/>
       </div>
     </div>