Previously, Always rendering whole blocks and it could possibly cause performance issue.
Please see [1]. It occurred at 20ms rendering time which over 16ms for 60 FPS.
It may cause a bad-user experience.

Therefore, Separate rendering layers to reduce rendering time.
The background tiles are rendered only once when the canvas is initialized.
Also, Stacked blocks are not required to render not every time. So add "shouldUpdateStackedLayer" to use checking update the stacked block.

So, Update blocks are only control-blocks.

Note here [2], the rendering time reduced to 13~6ms.

[1] https://user-images.githubusercontent.com/18409763/110893398-40e42480-8339-11eb-928e-fb41a4ca7020.png

[2] https://user-images.githubusercontent.com/18409763/110893887-34140080-833a-11eb-878f-7156b6fb81e7.png