@WoochanLee 2021-03-20
Please check this bug.
An afterimage remains when using the shift function.
<img width="606" alt="스크린샷 2021-03-20 오후 5 08 10" src="https://user-images.githubusercontent.com/25609385/111863448-00228600-899f-11eb-872c-03851ae502c3.png">