@WoochanLee 2021-03-20
Don't merge before has two approved.
I will change branch rules.